Struct bit_matrix::matrix::BitMatrix

pub struct BitMatrix {
    // some fields omitted
}

A matrix of bits.

Methods

impl BitMatrix

fn new(rows: usize, row_bits: usize) -> Self

Create a new BitMatrix with specific numbers of bits in columns and rows.

fn size(&self) -> (usize, usize)

Returns the matrix's size as (rows, columns).

fn set(&mut self, row: usize, col: usize, enabled: bool)

Sets the value of a bit.

Panics

Panics if (row, col) is out of bounds.

fn grow(&mut self, num_rows: usize, value: bool)

Grows the matrix in-place, adding num_rows rows filled with value.

fn sub_matrix(&self, range: Range<usize>) -> BitSubMatrix

Returns a slice of the matrix's rows.

fn split_at(&self, row: usize) -> (BitSubMatrix, &BitVecSlice, BitSubMatrixMut)

Given a row's index, returns a slice of all rows above that row, a reference to said row, and a slice of all rows below.

Functionally equivalent to (self.sub_matrix(0..row), &self[row], self.sub_matrix(row..self.num_rows())).

fn split_at_mut(&mut self, row: usize) -> (BitSubMatrixMut, &mut BitVecSlice, BitSubMatrixMut)

Given a row's index, returns a slice of all rows above that row, a reference to said row, and a slice of all rows below.

fn iter_row(&self, row: usize) -> Iter

Iterate over bits in the specified row.

fn transitive_closure(&mut self)

Computes the transitive closure of the binary relation represented by the matrix.

Uses the Warshall's algorithm.